// types/threads-responses.interface.ts
import { ProfileResponsesType } from './profile-responses.interface';

export interface ThreadsResponsesType {
  id: string;
  title: string;
  body: string;
  category: string;
  createdAt: string;
  ownerId: string;
  upVotesBy: string[];
  downVotesBy: string[];
  totalComments: number;
}

export interface ThreadsWithUserType {
  id: string;
  title: string;
  body: string;
  category: string;
  createdAt: string;
  ownerId: string;
  upVotesBy: string[];
  downVotesBy: string[];
  totalComments: number;
  user?: ProfileResponsesType;
}